requirements.txt
torch>=2.0.0
numpy
pandas
h5py
scikit-learn
requests
pyyaml
matplotlib
tensorboard
